Remove unknown attribute errors to ensure forward-compatibility
errAttributeUnknown was removed to make sure the library works on new kernels
when new Conntrack Netlink attributes are introduced. This recently occurred when
CTA_STATS_GLOBAL_MAX_ENTRIES was added in kernel 4.18. When a message containing
said attribute was encountered, the query would fail.

This patch ignores unknown and deprecated attributes.