conn: Add option to zero all accounting counters for Dump*
Apart from `IPCTNL_MSG_CT_GET` there's also `IPCTNL_MSG_CT_GET_CTRZERO`
which zeroes all the accounting counters that netfilter maintains.

Add a new struct to opt-in to zeroing the counters if need be.
The default behaviour remains the same, but all calls to
`Dump` and `DumpFilter` require a minor change to `Dump(nil)` and
`DumpFilter(nil)` when no options are used.