event: correctly parse TCP DESTROY events on kernels 5.11 and later
Linux commit 86d21fc74745 ("netfilter: ctnetlink: add timeout and protoinfo to
destroy events") added protoinfo to destroy events, but in doing so, changed
the logic used to assemble the ProtoInfoTCP payload, causing TCP flags to not be
sent as attributes.

This invalidated a long-standing assumption in the library that ProtoInfoTCP
would always contain at least 3 attributes. This commit relaxes the parser and
fixes the error returns of the SCTP and DCCP families.

TestConnListen now ranges over TCP, UDP, SCTP and DCCP families to verify if all
of NEW/UPDATE/DESTROY events can be parsed for all families.