Fix BenchmarkFlowUnmarshal
The benchmark was not correct as all iterations except for the first one
were no-ops. The netlink.AttributeDecoder cannot be reused multiple
times. The existing code was claiming to make a copy of it for each
iteration, but mustDecodeAttributes actually returns a pointer, so it
needs to be dereferenced to be copied.

Signed-off-by: Antonin Bas <abas@vmware.com>