Use netip.Addr instead of net.IP
Using the net/netip package instead of the net package can help reduce
the memory footprint of the library and help reduce the number of
heap allocations.

This is a breaking change for consumers of the library as exported types
are updated to use fields of type netip.Addr instead of net.IP.

We also remove the depedency on go-cmp and use assert.Equal instead in
tests.

Fixes #35

Signed-off-by: Antonin Bas <abas@vmware.com>