Fix Conn.Get to correctly set IP family for partial IPv6 filters
Conn.Get supports filters with only one of TupleOrig or TupleReply defined.
However, when using such a partial filter with IPv6 addresses, the function
failed to set the correct IP family (ProtoIPv6) in the netfilter header,
resulting in the error:

netfilter query: netlink receive: invalid argument

This change updates the condition to set the protocol family to IPv6 if
either TupleOrig or TupleReply uses an IPv6 address.

Signed-off-by: Milan Lenco <milan@zededa.com>