Add support for zone filtering for Dump and Flush
Support for the CTA_ZONE attribute for dump and flush requests was added
to the kernel in 6.8:
https://github.com/torvalds/linux/commit/eff3c558bb7e61c41b53e4c8130e514a5a4df9ba

We add an optional `Zone` field to the `Filter` struct in order to
expose this support. In the case of Dump, for clients who are only
interested in flows from a specific zone, this can be much more
efficient than dumping all flows, unmarshalling them, and filtering
based on the unmarshalled zone value, especially considering that the
library has no support for partial unmarshalling.

For older kernels, the attribute will be ignored. While we do not
include a version check in the implementation of the Flush / Dump
methods, we do make this requirement clear in their documentation.

Fixes #23

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>